+++ monolib-dart/lib/src/stream/on_stream_of_list_of_int.dart
@@ -1,5 +1,7 @@
 import 'dart:convert';
 
+import 'line_splitter_converter.dart';
+
 extension OnStreamOfListOfInt on Stream<List<int>> {
   Future<String> readLine() {
     return cast<List<int>>()
@@ -10,7 +12,6 @@ extension OnStreamOfListOfInt on Stream<List<int>> {
 
   Stream<String> utf8DecodeAndLineSplit() {
     return cast<List<int>>()
-        .transform(utf8.decoder)
-        .transform(const LineSplitter());
+        .transform(utf8.decoder.fuse(const LineSplitterConverter()));
   }
 }